+++ relay-spans/src/span.rs
@@ -77,6 +77,7 @@ pub fn otel_to_sentry_span(otel_span: OtelSpan) -> EventSpan {
         name,
         attributes,
         status,
+        kind,
         ..
     } = otel_span;
 
@@ -92,6 +93,9 @@ pub fn otel_to_sentry_span(otel_span: OtelSpan) -> EventSpan {
     };
 
     let mut op = None;
+    let mut description = name;
+    let mut http_method = None;
+    let mut http_route = None;
     let mut http_status_code = None;
     let mut grpc_status_code = None;
     for attribute in attributes.into_iter() {
@@ -103,6 +107,23 @@ pub fn otel_to_sentry_span(otel_span: OtelSpan) -> EventSpan {
             };
             if key == "sentry.op" {
                 op = otel_value_to_string(value);
+            } else if key.starts_with("db") {
+                op = op.or(Some("db".to_string()));
+                if key == "db.statement" {
+                    if let Some(statement) = otel_value_to_string(value) {
+                        description = statement;
+                    }
+                }
+            } else if key == "http.method" || key == "request.method" {
+                let http_op = match kind {
+                    2 => "http.server",
+                    3 => "http.client",
+                    _ => "http",
+                };
+                op = op.or(Some(http_op.to_string()));
+                http_method = otel_value_to_string(value);
+            } else if key == "http.route" || key == "url.path" {
+                http_route = otel_value_to_string(value);
             } else if key.contains("exclusive_time_ns") {
                 let value = match value {
                     OtelValue::IntValue(v) => v as f64,
@@ -148,10 +169,14 @@ pub fn otel_to_sentry_span(otel_span: OtelSpan) -> EventSpan {
     // TODO: This is wrong, a segment could still have a parent in the trace.
     let is_segment = parent_span_id.is_empty().into();
 
+    if let (Some(http_method), Some(http_route)) = (http_method, http_route) {
+        description = format!("{} {}", http_method, http_route);
+    }
+
     EventSpan {
         op: op.into(),
+        description: description.into(),
         data: SpanData::from_value(Annotated::new(data.into())),
-        description: name.into(),
         exclusive_time: exclusive_time_ms.into(),
         parent_span_id: SpanId(parent_span_id).into(),
         segment_id,
@@ -286,5 +311,81 @@ mod tests {
         let otel_span: OtelSpan = serde_json::from_str(json).unwrap();
         let event_span: EventSpan = otel_to_sentry_span(otel_span);
         assert_eq!(event_span.exclusive_time, Annotated::new(0.0788));
+    }
+
+    #[test]
+    fn parse_span_with_db_attributes() {
+        let json = r#"{
+            "traceId": "89143b0763095bd9c9955e8175d1fb23",
+            "spanId": "e342abb1214ca181",
+            "parentSpanId": "0c7a7dea069bf5a6",
+            "name": "database query",
+            "kind": 3,
+            "startTimeUnixNano": 1697620454980000000,
+            "endTimeUnixNano": 1697620454980078800,
+            "attributes": [
+                {
+                    "key" : "db.name",
+                    "value": {
+                        "stringValue": "database"
+                    }
+                },
+                {
+                    "key" : "db.type",
+                    "value": {
+                        "stringValue": "sql"
+                    }
+                },
+                {
+                    "key" : "db.statement",
+                    "value": {
+                        "stringValue": "SELECT \"table\".\"col\" FROM \"table\" WHERE \"table\".\"col\" = %s"
+                    }
+                }
+            ]
+        }"#;
+        let otel_span: OtelSpan = serde_json::from_str(json).unwrap();
+        let event_span: EventSpan = otel_to_sentry_span(otel_span);
+        assert_eq!(event_span.op, Annotated::new("db".into()));
+        assert_eq!(
+            event_span.description,
+            Annotated::new(
+                "SELECT \"table\".\"col\" FROM \"table\" WHERE \"table\".\"col\" = %s".into()
+            )
+        );
+    }
+
+    #[test]
+    fn parse_span_with_http_attributes() {
+        let json = r#"{
+            "traceId": "89143b0763095bd9c9955e8175d1fb23",
+            "spanId": "e342abb1214ca181",
+            "parentSpanId": "0c7a7dea069bf5a6",
+            "name": "http client request",
+            "kind": 3,
+            "startTimeUnixNano": 1697620454980000000,
+            "endTimeUnixNano": 1697620454980078800,
+            "attributes": [
+                {
+                    "key" : "http.request.method",
+                    "value": {
+                        "stringValue": "GET"
+                    }
+                },
+                {
+                    "key" : "url.path",
+                    "value": {
+                        "stringValue": "/api/search?q=foobar"
+                    }
+                }
+            ]
+        }"#;
+        let otel_span: OtelSpan = serde_json::from_str(json).unwrap();
+        let event_span: EventSpan = otel_to_sentry_span(otel_span);
+        assert_eq!(event_span.op, Annotated::new("http.client".into()));
+        assert_eq!(
+            event_span.description,
+            Annotated::new("GET /api/search?q=foobar".into())
+        );
     }
 }